refactor: 문자열, 숫자 분리하는 부분을 별도의 클래스로 정의
